Created project "Chip8 Interpreter"
This project is going to be an interpreter for the Chip8 interpreted language
Chip8 was developed in the mid 1970s by Joseph Weisbecker, and was
originally used on the COSMAC VIP and the Telmac 1800. I was interested
in seeing how complex that this would be to create, so this is my shot at it.