Setup basic interpreter layout
So Chip8 runs with only 4k memory, though only 2k was really ever used.
Chip8 also has 16 general purpose registers (V), though registers[0xF] is special
because it contains information when a pixel overwrites another pixel, or when
an integer over/underflows. It aslo has a stack of 16 bit addresses with a stack
pointer, and two 8 bit timers, one for delays and one for sound. I decided to
have the program counter (PC), saveAddress (I), and the stack pointer (SP) all
be actual pointers, so that I can challenge myself with some pointer arithematic
though not using pointers would be faster, since pointers are (in my case)
64-bit numbers. The chip8 also has a display resolution of 64x32 pixels. I also
added a draw flag so that the window only updates when the chip8 draws something.